+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/models/ParallelTransferOptions.java
@@ -16,10 +16,9 @@ import com.azure.storage.common.implementation.StorageImplUtils;
 public final class ParallelTransferOptions {
 
     private final Integer blockSize;
-    private final Integer numBuffers;
+    private final Integer maxConcurrency;
     private final ProgressReceiver progressReceiver;
     private final Integer maxSingleUploadSize;
-    private final Integer maxConcurrency;
 
     /**
      * Creates a new {@link ParallelTransferOptions} with default parameters applied.
@@ -30,14 +29,14 @@ public final class ParallelTransferOptions {
      * individual call will send more data and will therefore take longer. This parameter also determines the size
      * that each buffer uses when buffering is required and consequently amount of memory consumed by such methods may
      * be up to blockSize * numBuffers.
-     * @param numBuffers For buffered upload only, the number of buffers is the maximum number of buffers this method
-     * should allocate. Memory will be allocated lazily as needed. Must be at least two. Typically, the larger the
+     * @param maxConcurrency For buffered upload only, the number of buffers is the maximum number of buffers this
+     * method should allocate. Memory will be allocated lazily as needed. Must be at least two. Typically, the larger the
      * number of buffers, the more parallel, and thus faster, the upload portion  of this operation will be.
      * The amount of memory consumed by methods using this value may be up to blockSize * numBuffers.
      * @param progressReceiver {@link ProgressReceiver}
      */
-    public ParallelTransferOptions(Integer blockSize, Integer numBuffers, ProgressReceiver progressReceiver) {
-        this(blockSize, numBuffers, progressReceiver, null);
+    public ParallelTransferOptions(Integer blockSize, Integer maxConcurrency, ProgressReceiver progressReceiver) {
+        this(blockSize, maxConcurrency, progressReceiver, null);
     }
 
     /**
@@ -48,11 +47,15 @@ public final class ParallelTransferOptions {
      * of requests that need to be made. If block size is large, upload will make fewer network calls, but each
      * individual call will send more data and will therefore take longer. This parameter also determines the size
      * that each buffer uses when buffering is required and consequently amount of memory consumed by such methods may
-     * be up to blockSize * numBuffers.
-     * @param numBuffers For buffered upload only, the number of buffers is the maximum number of buffers this method
-     * should allocate. Memory will be allocated lazily as needed. Must be at least two. Typically, the larger the
-     * number of buffers, the more parallel, and thus faster, the upload portion  of this operation will be.
-     * The amount of memory consumed by methods using this value may be up to blockSize * numBuffers.
+     * be up to blockSize * maxConcurrency.
+     * @param maxConcurrency The maximum number of parallel requests that will be issued at any given time as a part of
+     * a single parallel transfer. This value applies per api. For example, if two calls to uploadFromFile are made at
+     * the same time, and each specifies a maxConcurrency of 5, there may be up to 10 outstanding, concurrent requests,
+     * up to 5 for each of the upload operations. For buffered uploads only, this value is the maximum number of buffers
+     * to be allocated as part of the transfer. In those cases, memory will be allocated lazily as needed and this value
+     * must be at least two. The amount of memory consumed by methods which buffer may be up to
+     * blockSize * maxConcurrency. In general, upload methods which do not accept a length parameter must perform some
+     * buffering.
      * @param progressReceiver {@link ProgressReceiver}
      * @param maxSingleUploadSize If the size of the data is less than or equal to this value, it will be uploaded in a
      * single put rather than broken up into chunks. If the data is uploaded in a single shot, the block size will be
@@ -62,45 +65,22 @@ public final class ParallelTransferOptions {
      * service accepts for uploading in a single requests and is represented by
      * {@link BlockBlobAsyncClient#MAX_UPLOAD_BLOB_BYTES}.
      */
-    public ParallelTransferOptions(Integer blockSize, Integer numBuffers, ProgressReceiver progressReceiver,
+    public ParallelTransferOptions(Integer blockSize, Integer maxConcurrency, ProgressReceiver progressReceiver,
             Integer maxSingleUploadSize) {
-        this(blockSize, numBuffers, progressReceiver, maxSingleUploadSize, null);
-    }
-
-    /**
-     * Creates a new {@link ParallelTransferOptions} with default parameters applied.
-     *
-     * @param blockSize The block size.
-     * For upload, The block size is the size of each block that will be staged. This value also determines the number
-     * of requests that need to be made. If block size is large, upload will make fewer network calls, but each
-     * individual call will send more data and will therefore take longer. This parameter also determines the size
-     * that each buffer uses when buffering is required and consequently amount of memory consumed by such methods may
-     * be up to blockSize * numBuffers.
-     * @param numBuffers For buffered upload only, the number of buffers is the maximum number of buffers this method
-     * should allocate. Memory will be allocated lazily as needed. Must be at least two. Typically, the larger the
-     * number of buffers, the more parallel, and thus faster, the upload portion  of this operation will be.
-     * The amount of memory consumed by methods using this value may be up to blockSize * numBuffers.
-     * @param progressReceiver {@link ProgressReceiver}
-     * @param maxSingleUploadSize If the size of the data is less than or equal to this value, it will be uploaded in a
-     * single put rather than broken up into chunks. If the data is uploaded in a single shot, the block size will be
-     * ignored. Some constraints to consider are that more requests cost more, but several small or mid-sized requests
-     * may sometimes perform better. In the case of buffered upload, up to this amount of data may be buffered before
-     * any data is sent. Must be greater than 0. May be null to accept default behavior, which is the maximum value the
-     * service accepts for uploading in a single requests and is represented by
-     * {@link BlockBlobAsyncClient#MAX_UPLOAD_BLOB_BYTES}.
-     * @param maxConcurrency The maximum number of parallel requests that will be issued at any given time.
-     */
-    public ParallelTransferOptions(Integer blockSize, Integer numBuffers, ProgressReceiver progressReceiver,
-        Integer maxSingleUploadSize, Integer maxConcurrency) {
         if (blockSize != null) {
             StorageImplUtils.assertInBounds("blockSize", blockSize, 1, BlockBlobAsyncClient.MAX_STAGE_BLOCK_BYTES);
         }
         this.blockSize = blockSize;
 
-        if (numBuffers != null) {
-            StorageImplUtils.assertInBounds("numBuffers", numBuffers, 2, Integer.MAX_VALUE);
+        /*
+        In buffered upload cases, maxConcurrency must be at least 2 as it also indicates the number of buffers we will
+        allocate. That minimum is validated in the UploadBufferPool. Because this value is also used in file transfers
+        which do not buffer, we only check the absolute minimum for this value here.
+         */
+        if (maxConcurrency != null) {
+            StorageImplUtils.assertInBounds("maxConcurrency", maxConcurrency, 1, Integer.MAX_VALUE);
         }
-        this.numBuffers = numBuffers;
+        this.maxConcurrency = maxConcurrency;
         this.progressReceiver = progressReceiver;
 
         if (maxSingleUploadSize != null) {
@@ -108,8 +88,6 @@ public final class ParallelTransferOptions {
                 BlockBlobAsyncClient.MAX_UPLOAD_BLOB_BYTES);
         }
         this.maxSingleUploadSize = maxSingleUploadSize;
-
-        this.maxConcurrency = maxConcurrency;
     }
 
     /**
@@ -123,9 +101,10 @@ public final class ParallelTransferOptions {
     /**
      * Gets the number of buffers being used for a transfer operation.
      * @return The number of buffers.
+     * @deprecated Use {@link #getMaxConcurrency()}
      */
     public Integer getNumBuffers() {
-        return this.numBuffers;
+        return this.maxConcurrency;
     }
 
     /**